Refactor utility/decomp headers
