+++ LEGO1/realtime/matrix.cpp
@@ -1,7 +1,7 @@
 
 #include "matrix.h"
 
-#include "../decomp.h"
+#include "decomp.h"
 #include "math.h"
 
 #include <memory.h>
