+++ util/compat.h
@@ -1,5 +1,5 @@
-#ifndef ISLECOMPAT_H
-#define ISLECOMPAT_H
+#ifndef COMPAT_H
+#define COMPAT_H
 
 // Various macros to enable compiling with other/newer compilers.
 
@@ -11,26 +11,12 @@
 #define COMPAT_CONST
 #endif
 
-// DIsable "nonstandard extension used : 'bool'" warning spam
-#pragma warning(disable : 4237)
-
 // Disable "identifier was truncated to '255' characters" warning.
 // Impossible to avoid this if using STL map or set.
 // This removes most (but not all) occurrences of the warning.
 #pragma warning(disable : 4786)
 
 #define MSVC420_VERSION 1020
-
-// STL compatibility.
-#if defined(_MSC_VER) && _MSC_VER <= MSVC420_VERSION
-#include "mxstl.h"
-#else
-#include <algorithm>
-#include <list>
-#include <set>
-using std::list;
-using std::set;
-#endif
 
 // We use `override` so newer compilers can tell us our vtables are valid,
 // however this keyword was added in C++11, so we define it as empty for
@@ -39,4 +25,4 @@ using std::set;
 #define override
 #endif
 
-#endif // ISLECOMPAT_H
+#endif // COMPAT_H
